requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-SocketIO==5.3.6
Flask-HTTPAuth==4.4.0
PyJWT==2.8.0
redis==5.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23
alembic==1.12.1
email-validator==2.1.0.post1
python-socketio==5.10.0
python-engineio==4.8.0
simple-websocket==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9 